Move heading-related logics to core package
