+++ packages/core/src/index.ts
@@ -3,3 +3,4 @@ export * from './constants';
 export * from './utils';
 export * from './diff';
 export * from './steps';
+export * from './heading';
